backend/requirements.txt
flask
# flask_sqlalchemy
flask_httpauth
# mysql.connector>=2.2.9
mysql-connector-python
requests
